rpc: add new RPCConnection class
It's purpose is to have a universal way to create clients,
and keep track of current connection settings.

A side effect is that we read config files inside each rpc,
and failing initialization doesn't crash the entire app.

Reasoning behind not crashing on startup is that an rpc
connection can fail at any time, and sidesail should know
immediately when a connection is bad! The user might shut
down their node, it might crash etc. Sidesail will probably
stay open for days at a time. Most people rarely close their
applications.

It's a bit messy right now, hoping to make it cleaner in a later
iteration.