tabs: add testchain rpc tab page
Identical to the ethereum rpc page, but for testchain!

I did not expect this, but it ends up creating a good UX.

On trying out a command, it's immediately clear to the user that
bitcoind is running behind the scenes.

Up top of the rpc-list are sidechain-specific rpcs, that anyone familiar
with bitcoind will not recognize.

Instantly gives the user the correct mental model, lets them know
there's a cli available (if they so prefer), and removes the magic.