app+home: move daemon error handling to home page
With this change, we do an optimistic approach, and assume the
connection already exists, or will resolve pretty fast. The user
will be presented with a loading spinner for each individual chain
during the daemon initialization phase.

We also assume the user is pretty competent, and allow them to tweak
connection settings by going to the connection page. In case they're
running some sort of custom setup.

We don't block the user from opening the UI if the connection fails,
but assume they understand nothing will work when a gigantic red
error is presented in a popup, and is visible in the side navigation.

If the daemons are already running in the background, no popup will be
shown, and the app will launch very fast.