drivechain_client: add balances view
Uses a dummy btc price of 50k USD. Let me know if there's stuff I should get from sail_ui instead.

I was also unable to get a bold font with a sail-styled `Text` or `SailText` so I had to resort to `RichText` for some of the labels, which is quite verbose.